+++ www/src/MDX/Ol.tsx
@@ -25,11 +25,11 @@
  */
 
 import React, { FC } from 'react'
-import { List } from '@looker/components'
+import { OrderedList } from '@looker/components'
 import { maxTextWidth } from './styles'
 
-export const Ol: FC<{}> = (props) => (
-  <List
+const Ol: FC<{}> = (props) => (
+  <OrderedList
     lineHeight="medium"
     mb="medium"
     ml="large"